ExpenseTransactionDAO and ExpenseReconciler This is restructure of the code to adopt 2 design patterns: 1. ExpenseTransactionDAO ExpenseTransactionDAO adopts the Data Access Object pattern to create an additional layer of abstraction to the retrieval of Expense Transaction objects from the sqLite database. With this layer, the database query will be encapsulated, together with the existing ExpenseTransactionMapper. 2. ExpenseReconciler ExpenseReconciler adopts the Service abstraction pattern. It create an layer of abstraction of how the data retrieved from the CSV and the database are used for reconcilation. It allows for expansion in the future without affects the main application code.
A first attempt to adopt the Data Access Object pattern and Service pattern
So now the retrieval of the expense transaction is being encapsulated in the data object access class. So the knowledge of access the database is hidden from the application class. A new service class ExpenseReconciler is create to perform a business service of reconciling the given data from the CSV file and the database