Rename database to persistence and move CsvParser there
Rename the database package to a more generic naming - persistence package. With this change, handling of all persistence related operations will be stored in the package. Along with this change, since CsvParser reads CSV file, which is a form of persistant storage, the class is moved to the persistence package.