Bugfix: Maven output jar doesn't work with dependency complain
- There is an error after Maven builds the Java jar files

Exception in thread "main" java.lang.NoClassDefFoundError: org/apache/logging/log4j/LogManager
  at cracking_the_coding_interview.UniqueCharacter.<clinit>(UniqueCharacter.java:8)
Caused by: java.lang.ClassNotFoundException: org.apache.logging.log4j.LogManager
  at java.base/jdk.internal.loader.BuiltinClassLoader.loadClass(BuiltinClassLoader.java:583)
  at java.base/jdk.internal.loader.ClassLoaders$AppClassLoader.loadClass(ClassLoaders.java:178)
  at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:521)

This is found out to be an dependency issue. The root cause is found that the jar files can't reference the required dependencies library, and the solution is to move all the dependencies to a specified folder stated in the manifest file.

There is also an small enhancement made to move all the unecessary resource files used previously to the test folder. They are meant to be deleted after the program is stable.