Update the top layer application naming. By updating and enforcing better coding practices into the code, the readability and maintainability of the code is increased. (#13)
Sidenote: There has been thought of using the Main class as the Inversion of Control container. However, there are 2 reaons against it. First, the service ExpenseReportReader needs the database filename to be initialised. At system startup, before reading any parameter, this is not possible. Given that the task of processing the parameter has been passed down the layer to the ExpenseAccountant. One way around this is to use the ExpenseAccountant as the IoC container, but this defeat the purpose of having all the dependencies injected to ExpenseAccountant so that this ExpenseAccountant will be a Service class as well. Moreover, after studying the life cylce of each Service class, there isn't a need to maintain the state of the Service. In the end, this decision is discarded.