New base module and new dependency Apache Commons Lang3 library
A new module called base is added to the application. This new base module will contain cross-modules classes and enums. These classes and enums should be applicable for all modules. They will contain non-functional code, such as utility classes, collections of constants. The first enum added is to store application-wide non-functional String constant.

After a long period of resisted temptation, Apache Commons Lang3 library is added. There are a lot of useful methods that are beneficial, and there is no point reinventing the wheel and it is not the objective of this application as well. Currently, StringUtils is used.