Reorganise the Maven dependencies
Split the existing root Maven project dependencies. After the split, dependency for each child Maven project will contain the minimum dependency required to work on its own. The test dependency is also declared for each child Maven project so that the jar are not included in the exporting jar.

The task to copy library dependencies is also shift to cli child Maven project.